Fix size of length param

+++ Inc/spi_cam.h
@@ -23,7 +23,7 @@ typedef enum {
 
 OV5642_StatusTypeDef OV5642_Init(I2C_HandleTypeDef *hi2c, SPI_HandleTypeDef *hspi, GPIO_TypeDef *cs_gpio_port, uint16_t cs_gpio_pin);
 
-OV5642_StatusTypeDef OV5642_Take_Picture(uint8_t frame_buffer[], uint8_t *length);
+OV5642_StatusTypeDef OV5642_Take_Picture(uint8_t frame_buffer[], uint32_t *length);
 
 
 void Error_Handler(void);
